fixed usersave method add auth to journal endpoint

+++ src/main/java/com/nketu/journal/service/JournalEntryService.java
@@ -3,7 +3,6 @@ package com.nketu.journal.service;
 import com.nketu.journal.entity.JournalEntry;
 import com.nketu.journal.entity.User;
 import com.nketu.journal.repository.JournalEntryRepository;
-import com.nketu.journal.repository.UserRepository;
 import org.bson.types.ObjectId;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,7 +35,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception e){
             log.error("exception",e);
             throw new RuntimeException("An error occurred while saving th entry.",e);
@@ -51,11 +50,27 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed=false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed= user.getJournalEntries().removeIf(x->x.getId().equals(id));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting "+e);
+        }
+        return removed;
+
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
 
 }
